Terminado de hacer los components
